docs: Added cube comments
