+++ rushcargo-app/src/model/app.rs
@@ -147,84 +147,4 @@ impl App {
             }
         }
     }
-
-    fn update_cube(&mut self) {
-        if let Some(title) = self.title.as_mut() {
-                let mut counter = 0;
-                for dot in title.cube.dot.iter() {
-                let mut x = dot.x as f64;
-                let mut y = dot.y as f64;
-                let mut z = dot.z as f64;
-                
-                let x1 = &mut title.cube.rot_dot[counter].x;
-                let y1 = &mut title.cube.rot_dot[counter].y;
-                let z1 = &mut title.cube.rot_dot[counter].z;
-
-                (x, y, z) = Self::rotate_x(title.cube.A, x as f64, y as f64, z as f64);
-                (x, y, z) = Self::rotate_y(title.cube.B, x as f64, y as f64, z as f64);
-                (x, y, z) = Self::rotate_z(title.cube.C, x as f64, y as f64, z as f64);
-                //title.cube.rot_dot[counter].x = Self::calculate_x(title.cube.A, title.cube.B, title.cube.C, x, y, z);
-                //title.cube.rot_dot[counter].y = Self::calculate_y(title.cube.A, title.cube.B, title.cube.C, x, y, z);
-                //title.cube.rot_dot[counter].z = Self::calculate_z(title.cube.A, title.cube.B, title.cube.C, x, y, z);
-                (*x1, *y1, *z1) = (x, y, z);
-
-                counter += 1;
-            }
-
-            //println!("{}", title.cube.rot_dot.x);
-            //println!("{}", title.cube.rot_dot.y);
-            //println!("{}", title.cube.rot_dot.z);
-            //println!();
-
-            title.cube.A += 0.1;
-            title.cube.B += 0.1;
-            //title.cube.C += 0.1;
-
-        }
-    }
-    
-    /*fn calculate_x(A: f64, B: f64, C: f64, i: i8, j: i8, k: i8) -> f64 {
-        j as f64 * A.sin() * B.sin() * C.cos() - k as f64 * A.cos() * B.sin() * C.cos() +
-        j as f64 * A.cos() * C.sin() + k as f64 * A.sin() * C.sin() + i as f64 * B.cos() * C.cos()
-    }*/
-    
-    fn calculate_x(A: f64, B: f64, C: f64, i: i8, j: i8, k: i8) -> f64 {
-        i as f64 * C.cos() * B.cos() +
-        j as f64 * (C.cos() * B.sin() * A.sin() - C.sin() * A.cos()) +
-        k as f64 * (C.cos() * B.sin() * A.cos() + C.sin() * A.sin())
-    }
-
-    fn rotate_x(ang: f64, i: f64, j: f64, k: f64) -> (f64, f64, f64) {
-        (
-            i,
-            j * ang.cos() + k * ang.sin(),
-            k * ang.cos() - j * ang.sin()
-        )
-    }
-    
-    fn rotate_y(ang: f64, i: f64, j: f64, k: f64) -> (f64, f64, f64) {
-        (
-            i * ang.cos() - k * ang.sin(),
-            j,
-            i * ang.sin() + k.cos()
-        )
-    }
-
-    fn rotate_z(ang: f64, i: f64, j: f64, k: f64) -> (f64, f64, f64) {
-        (
-            i * ang.cos() + j * ang.sin(),
-            j * ang.cos() - i * ang.sin(),
-            k
-        )
-    }
-
-    fn calculate_y(A: f64, B: f64, C: f64, i: i8, j: i8, k: i8) -> f64 {
-        j as f64 * A.cos() * C.cos() + k as f64 * A.sin() * C.cos() -
-        j as f64 * A.sin() * B.sin() + C.sin() + k as f64 * A.cos() * B.sin() * C.sin() -
-        i as f64 * B.cos() * C.sin()
-    }
-
-    fn calculate_z(A: f64, B: f64, C: f64, i: i8, j: i8, k: i8) -> f64 {
-        k as f64 * A.cos() * B.cos() - j as f64 * A.sin() * B.cos() + i as f64 * B.sin()
-    }
 }
